+++ cpp/subprojects/common/src/mlrl/common/rule_refinement/prediction_partial.cpp
@@ -137,7 +137,7 @@ std::unique_ptr<IStatisticsSubset> PartialPrediction<ScoreType>::createStatistic
 
 template<typename ScoreType>
 std::unique_ptr<IStatisticsSubset> PartialPrediction<ScoreType>::createStatisticsSubset(
-  const IStatistics& statistics, const DenseWeightVector<uint32>& weights) const {
+  const IStatistics& statistics, const DenseWeightVector<uint16>& weights) const {
     return statistics.createSubset(indexVector_, weights);
 }
 
@@ -161,7 +161,7 @@ std::unique_ptr<IStatisticsSubset> PartialPrediction<ScoreType>::createStatistic
 
 template<typename ScoreType>
 std::unique_ptr<IStatisticsSubset> PartialPrediction<ScoreType>::createStatisticsSubset(
-  const IStatistics& statistics, const OutOfSampleWeightVector<DenseWeightVector<uint32>>& weights) const {
+  const IStatistics& statistics, const OutOfSampleWeightVector<DenseWeightVector<uint16>>& weights) const {
     return statistics.createSubset(indexVector_, weights);
 }
